requirements.txt
amqp==5.1.1
arabic-reshaper==2.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
bleach==4.1.0
celery==5.2.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coverage==6.2
cryptography==37.0.4
cssselect2==0.6.0
cx-Oracle==8.3.0
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.5
Django==3.2.7
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-crispy-forms==1.14.0
django-environ==0.9.0
django-filter==21.1
django-hosts==5.1
django-multiselectfield==0.1.12
django-pagination==1.0.7
django-simple-pagination==1.3
django-timezone-field==5.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
entrypoints==0.3
Faker==11.1.0
filelock==3.7.1
fonttools==4.31.1
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
httpie==2.6.0
idna==3.3
importlib-metadata==4.8.2
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.3.2
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.4.0
kombu==5.2.4
lxml==4.9.1
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mysql-connector-python==8.0.28
mysqlclient==2.1.0
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
nose2==0.10.0
notebook==6.4.6
numpy==1.21.5
oscrypto==1.3.0
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.5.2
prometheus-client==0.12.0
prompt-toolkit==3.0.24
py4j==0.10.9.5
pycparser==2.21
pycryptodome==3.14.1
Pygments==2.10.0
pyHanko==0.13.2
pyhanko-certvalidator==0.19.5
PyJWT==1.7.1
pyparsing==3.0.6
PyPDF3==1.0.6
pyrsistent==0.18.0
PySocks==1.7.1
pyspark==3.3.0
python-bidi==0.4.2
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32==303
pywinpty==1.1.6
PyYAML==6.0
pyzmq==22.3.0
qrcode==7.3.1
qtconsole==5.2.2
QtPy==2.0.0
razorpay==1.3.0
redis==4.3.4
reportlab==3.6.2
requests==2.26.0
requests-toolbelt==0.9.1
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
svglib==1.4.1
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
threadpoolctl==3.0.0
tinycss2==1.1.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.7
vine==5.0.0
virtualenv==20.16.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
wikipedia==1.4.0
wrapt==1.14.1
xhtml2pdf==0.2.8
XlsxWriter==3.0.3
zipp==3.6.0
